# replace EntryNotFoundError on ErrEntryNotFound