I reconsidered my approach to `EntryNotFoundError` error. User always knows what key he uses, so we can delete `EntryNotFoundError` structure and replace it to `ErrEntryNotFound`


```
benchmark                                                old ns/op     new ns/op     delta
BenchmarkReadFromCacheNonExistentKeys/1-shards-12        185           181           -2.16%
BenchmarkReadFromCacheNonExistentKeys/512-shards-12      182           179           -1.65%
BenchmarkReadFromCacheNonExistentKeys/1024-shards-12     184           180           -2.17%
BenchmarkReadFromCacheNonExistentKeys/8192-shards-12     182           176           -3.30%

benchmark                                                old allocs     new allocs     delta
BenchmarkReadFromCacheNonExistentKeys/1-shards-12        1              0              -100.00%
BenchmarkReadFromCacheNonExistentKeys/512-shards-12      1              0              -100.00%
BenchmarkReadFromCacheNonExistentKeys/1024-shards-12     1              0              -100.00%
BenchmarkReadFromCacheNonExistentKeys/8192-shards-12     1              0              -100.00%

benchmark                                                old bytes     new bytes     delta
BenchmarkReadFromCacheNonExistentKeys/1-shards-12        24            8             -66.67%
BenchmarkReadFromCacheNonExistentKeys/512-shards-12      23            7             -69.57%
BenchmarkReadFromCacheNonExistentKeys/1024-shards-12     24            8             -66.67%
BenchmarkReadFromCacheNonExistentKeys/8192-shards-12     23            7             -69.57%
```